requirements.txt
pydantic-ai
python-dotenv
openai
requests
markdownify
sendgrid
sqlalchemy
pymysql